+++ lib/ui/groups/GroupDetailPanel.dart
@@ -19,7 +19,6 @@ import 'dart:typed_data';
 
 import 'package:expandable_page_view/expandable_page_view.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:neom/model/Analytics.dart';
 import 'package:neom/service/FlexUI.dart';
@@ -597,7 +596,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
           child: _buildImageHeader(),
         ),
         Padding(
-          padding: EdgeInsets.only(left: 24.0, right: 24.0, bottom: _hasGroupImage ? 72.0 : 24.0, top: _hasGroupImage ? 152.0 : 24.0),
+          padding: EdgeInsets.only(left: 24.0, right: 24.0, bottom: 56.0, top: _hasGroupImage ? 152.0 : 24.0),
           child: _buildGroupDetailsHeader(),
         )
       ]),
@@ -617,7 +616,6 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
   Widget _buildGroupDetailsHeader() {
     return Container(
       color: Styles().colors.surface,
-      // height: _groupDetailsHeight,
       child: Column(
         children: [
           _buildGroupInfo(),
@@ -1020,9 +1018,9 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
       children: [
         Expanded(
           child: RichText(textScaler: MediaQuery.of(context).textScaler,
-            text: TextSpan(text: Localization().getStringEx("","Your activity in the app is private. Please review the "), style:  Styles().textStyles.getTextStyle("widget.title.tiny"),
+            text: TextSpan(text: Localization().getStringEx("","Your activity in the app is private. Please review the "), style:  Styles().textStyles.getTextStyle("widget.title.dark.tiny"),
             children: [
-              TextSpan(text: Localization().getStringEx("", "Student Code."), style: Styles().textStyles.getTextStyle("widget.title.tiny")),  //recognizer: TapGestureRecognizer()..onTap = () => _onPrivacy()),
+              TextSpan(text: Localization().getStringEx("", "Student Code."), style: Styles().textStyles.getTextStyle("widget.title.dark.tiny")),  //recognizer: TapGestureRecognizer()..onTap = () => _onPrivacy()),
               // WidgetSpan(
               //     child: Padding(padding: EdgeInsets.symmetric(horizontal: 2), child: Styles().images.getImage('external-link', excludeFromSemantics: true)),
               // )
@@ -1218,7 +1216,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
     int leftAreaFlex = 0,
   }) {
     return Dialog(
-        backgroundColor: Styles().colors.fillColorPrimary,
+        backgroundColor: Styles().colors.background,
         child: StatefulBuilder(builder: (context, setStateEx) {
           return Padding(
               padding: EdgeInsets.all(16),
@@ -1231,9 +1229,9 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
                   Expanded(flex: leftAreaFlex, child: Container()),
                   Expanded(flex: negativeButtonFlex, child: RoundedButton(
                       label: StringUtils.ensureNotEmpty(negativeButtonLabel, defaultValue: Localization().getStringEx("panel.group_detail.button.back.title", "Back")),
-                      textStyle: Styles().textStyles.getTextStyle("widget.button.title.large"),
-                      borderColor: Styles().colors.surface,
-                      backgroundColor: Styles().colors.surface,
+                      textStyle: Styles().textStyles.getTextStyle("widget.button.light.title.large.fat"),
+                      borderColor: Styles().colors.fillColorSecondary,
+                      backgroundColor: Styles().colors.background,
                       padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
                       onTap: () {
                         Analytics().logAlert(text: confirmationTextMsg, selection: negativeButtonLabel);
@@ -1242,9 +1240,9 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
                   Container(width: 16),
                   Expanded(flex: positiveButtonFlex, child: RoundedButton(
                     label: positiveButtonLabel ?? '',
-                    textStyle: Styles().textStyles.getTextStyle("widget.button.title.large.fat"),
-                    borderColor: Styles().colors.surface,
-                    backgroundColor: Styles().colors.surface,
+                    textStyle: Styles().textStyles.getTextStyle("widget.button.light.title.large.fat"),
+                    borderColor: Styles().colors.fillColorSecondary,
+                    backgroundColor: Styles().colors.background,
                     padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
                     progress: _confirmationLoading,
                     onTap: () {
@@ -1286,7 +1284,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
                         label: Localization().getStringEx("panel.group_detail.button.group.about.title", "About this group"),//TBD localize
                         onTap: () {
                           Navigator.pop(context);
-                          setStateIfMounted(()=> _currentTab = _DetailTab.About);
+                          _onTab(_DetailTab.About);
                         })),
                 Visibility(
                     visible: _canEditGroup,
@@ -1354,7 +1352,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
                         })),
                 Visibility(visible: _canReportAbuse, child: RibbonButton(
                   leftIconKey: "report",
-                  label: Localization().getStringEx("panel.group.detail.post.button.report.students_dean.labe", "Report to Dean of Students"),
+                  label: Localization().getStringEx("panel.group.detail.post.button.report.students_dean.label", "Report to Dean of Students"),
                   onTap: () => _onTapReportAbuse(options: GroupPostReportAbuseOptions(reportToDeanOfStudents : true)   ),
                 )),
               ]));
@@ -1529,7 +1527,7 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> with TickerProvider
 
   void _onTapSettings(){
     Analytics().logSelect(target: "Group Settings", attributes: _group?.analyticsAttributes);
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupSettingsPanel(group: _group,))).then((exit){
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupSettingsPanel(group: _group, groupStats: _groupStats,))).then((exit){
       if(exit == true){
         Navigator.of(context).pop();
       }