+++ lib/ui/groups/GroupPostReportAbuse.dart
@@ -148,8 +148,8 @@ class _GroupPostReportAbusePanelState extends State<GroupPostReportAbusePanel> {
             RoundedButton(
               label: Localization().getStringEx("panel.group.detail.post.report_abuse.button.send.label", "Send"),
               hint: Localization().getStringEx("panel.group.detail.post.report_abuse.button.send.hint", "Tap to send report"),
-              textStyle: Styles().textStyles.getTextStyle("widget.button.title.large.fat"),
-              backgroundColor: Styles().colors.surface,
+              textStyle: Styles().textStyles.getTextStyle("widget.button.light.title.large.fat"),
+              backgroundColor: Styles().colors.background,
               borderColor: Styles().colors.fillColorSecondary,
               progress:  _sending,
               onTap: _onSend,